reorganize whole APP project
