+++ APP/.idea/gradle.xml
@@ -11,7 +11,7 @@
             <option value="$PROJECT_DIR$/app" />
           </set>
         </option>
-        <option name="useAutoImport" value="true" />
+        <option name="resolveModulePerSourceSet" value="false" />
       </GradleProjectSettings>
     </option>
   </component>
