+++ APP/app/src/main/java/passive_learner/urbrob/github/com/app/MainActivity.kt
@@ -1,0 +1,12 @@
+package passive_learner.urbrob.github.com.app
+
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
