+++ APP/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">passive-learner</string>
+    <string name="app_name">APP</string>
 </resources>
